Move strain and track data endpoints from gbrowser & indel finder to API
