+++ src/modules/site-v2/base/views/api/__init__.py
@@ -1,3 +1,5 @@
 from .gene          import api_gene_bp
 from .notifications import api_notifications_bp
+from .strains       import api_strains_bp
+from .tracks        import api_tracks_bp
 from .trait         import api_trait_bp
