+++ src/modules/site-v2/templates/tools/genome_browser/gbrowser.html
@@ -258,7 +258,7 @@ $(document).ready(function () {
   // Retrieves an object with two fields:
   //   - 'default':   the list of tracks that are not specific to any one strain
   //   - 'templates': the list of track templates to be filled out with strain data
-  fetch_json("{{ url_for('genome_browser.get_tracks') }}")
+  fetch_json("{{ url_for('api_tracks.get_tracks') }}")
     .then((data) => {
       set_trackset(data.default);
       set_templates(data.templates);
